# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Auth service",
        "title": "AuthService API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/user/deleteFollower/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint orqali foydalanuvchidan kuzatuvchini o'chirishingiz mumkin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Foydalanuvchidan kuzatuvchini o'chirish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token bilan birga keladi",
                        "name": "Athorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "O'chirilishi kerak bo'lgan kuzatuvchi IDsi",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kuzatuvchini muvaffaqiyatli o'chirdi",
                        "schema": {
                            "$ref": "#/definitions/users.Massage"
                        }
                    },
                    "400": {
                        "description": "Xatolik yuz berdi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/user/deleteUser{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint ma'lum bir foydalanuvchini ID orqali o'chirish uchun ishlatiladi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Foydalanuvchini o'chirish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Foydalanuvchi IDsi",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Foydalanuvchi muvaffaqiyatli o'chirildi",
                        "schema": {
                            "$ref": "#/definitions/users.Massage"
                        }
                    },
                    "400": {
                        "description": "Xatolik yuz berdi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/user/followers": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint orqali foydalanuvchining kuzatuvchilarini ro'yxatini olish mumkin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Foydalanuvchining kuzatuvchilarini olish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token bilan birga keladi",
                        "name": "Athorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Qaysi sahifani ko'rmoqchisiz? Default: 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Bir sahifada nechta element bo'lishi kerak? Default: 10",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Foydalanuvchining kuzatuvchilar ro'yxati",
                        "schema": {
                            "$ref": "#/definitions/users.GetaFollowersRes"
                        }
                    },
                    "400": {
                        "description": "Xatolik yuz berdi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/user/following": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint orqali ma'lum bir foydalanuvchining kimlarni kuzatayotganini olish mumkin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Foydalanuvchi kimlarni kuzatayotganini olish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token bilan birga keladi",
                        "name": "Athorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Sahifa raqami",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Sahifadagi elementlar soni",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kuzatadigan foydalanuvchilar ro'yxati",
                        "schema": {
                            "$ref": "#/definitions/users.GetaFollowingRes"
                        }
                    },
                    "400": {
                        "description": "Xatolik yuz berdi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/user/getFollower/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint orqali ma'lum bir kuzatuvchining ma'lumotlarini olish mumkin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Kuzatuvchi ma'lumotlarini olish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token bilan birga keladi",
                        "name": "Athorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Kuzatuvchi IDsi",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kuzatuvchi ma'lumotlari",
                        "schema": {
                            "$ref": "#/definitions/users.Follow"
                        }
                    },
                    "400": {
                        "description": "Xatolik yuz berdi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/user/getUsers": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint mavjud foydalanuvchilar ro'yxatini olish uchun ishlatiladi. Ro'yxat sahifalash orqali cheklangan.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Foydalanuvchilarni olish",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sahifa raqami, default: 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Har bir sahifadagi foydalanuvchilar soni, default: 10",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Foydalanuvchilar ro'yxati muvaffaqiyatli qaytarildi",
                        "schema": {
                            "$ref": "#/definitions/users.GetUserRes"
                        }
                    },
                    "400": {
                        "description": "Xatolik yuz berdi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/user/getUser{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint ma'lum bir foydalanuvchini ID orqali olish uchun ishlatiladi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Foydalanuvchini ID orqali olish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Foydalanuvchi IDsi",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Foydalanuvchi ma'lumotlari muvaffaqiyatli olindi",
                        "schema": {
                            "$ref": "#/definitions/users.User"
                        }
                    },
                    "400": {
                        "description": "Xatolik yuz berdi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/user/subscribe": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint orqali foydalanuvchi boshqa foydalanuvchini kuzatishi mumkin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Foydalanuvchini kuzatish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token bilan birga keladi",
                        "name": "Athorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Kuzatilayotgan foydalanuvchi IDsi",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kuzatish muvaffaqiyatli amalga oshirildi",
                        "schema": {
                            "$ref": "#/definitions/users.Massage"
                        }
                    },
                    "400": {
                        "description": "Xatolik yuz berdi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/users/deleteFollowing/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint orqali ma'lum bir foydalanuvchining kuzatayotgan boshqa foydalanuvchisini o'chirish mumkin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Foydalanuvchi kuzatayotgan foydalanuvchini o'chirish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token bilan birga keladi",
                        "name": "Athorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Kuzatishni to'xtatmoqchi bo'lgan foydalanuvchi ID'si",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "O'chirish muvaffaqiyatli amalga oshirildi",
                        "schema": {
                            "$ref": "#/definitions/users.Massage"
                        }
                    },
                    "400": {
                        "description": "Xatolik yuz berdi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/users/getFollowing/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint orqali ma'lum bir foydalanuvchining kimni kuzatayotganini ID orqali olish mumkin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Foydalanuvchi kimni kuzatayotganini ID orqali olish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token bilan birga keladi",
                        "name": "Athorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Kuzatmoqchi bo'lgan foydalanuvchi ID'si",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kuzatmoqchi bo'lgan foydalanuvchining ma'lumotlari",
                        "schema": {
                            "$ref": "#/definitions/users.Follow"
                        }
                    },
                    "400": {
                        "description": "Xatolik yuz berdi",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "users.Follow": {
            "type": "object",
            "properties": {
                "birth_day": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "users.GetaFollowersRes": {
            "type": "object",
            "properties": {
                "followers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/users.Followers"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                }
            }
        },
        "users.GetaFollowingRes": {
            "type": "object",
            "properties": {
                "following": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/users.Following"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                }
            }
        },
        "users.Massage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}